4pm push
Changes this push
- Fixed a bug that broke the entire placement system due to a incorrect object being called. (20mins)

- Adding UI telling the player the current mode. (Preview, Edit, Shopping and Placement). (10mins)

- Fixed bug with editing mode also breaking. Resolved with not soft locking the object selection to 1 object T_T (10mins)

- Added audio to the placement system (10mins) (Also crashed unity 👎)

- Fixed a bunch of little bugs with the hotbar and certain events. (20mins)

- Made UI in figma then imported it in. Then updated the playerstatics script (40mins)

Rest of today has been messing around with getting placements working better and also getting object data so buildings can be upgraded

Notes:
- The stupid script that is from the tutorial seems to not be a fan of placing objects anymore?! I don't know specifics as to why but placement does work eventually.

Resolved that issue above. The placement won't work when the mouse is currently over a object that isn't the ground so any foliage can block the mouse from placing. Idrk how to fix this as the layermask is also not giving two F's about the fact that there is a obviously large tree in the way :/